Move genValArg to testlib
