+++ plutus-core/testlib/PlutusCore/Generators.hs
@@ -4,6 +4,7 @@ module PlutusCore.Generators
     ( module Export
     ) where
 
+import PlutusCore.Generators.Internal.Builtin as Export
 import PlutusCore.Generators.Internal.Denotation as Export
 import PlutusCore.Generators.Internal.Entity as Export
 import PlutusCore.Generators.Internal.TypeEvalCheck as Export
